Add in redux
Added in the basics of the reux implementation.  Created the back bone for storing the data in the application.  Also modified a couple of the compilation rules.

I initially added in @ngrx for the angular redux implementation before changing my mind, so I'm not sure if there are artifacts in the package-lock.  Will probably regenerate that in the near future.