Color Square on Canvas
Tagged on a listener to the canvas for clicks to change the color of the square.  Also added a new basic input for color to hold the value for the square.  Ideally, that will get replaced with a color picker in the future.

The currenty canvas listener is for click - this might need to get broken down into mouse actions more.