Better Zoom
Tweaked a hell of a lot of logic in the canvas utility to cause zoom to follow the mouse a lot better.  Also raised the max allowed zoom up to 4x.  Will need to migrate the logic for determining zoom into the reducer in the future.

That logic is also a candidate for WA replacement.