Zoom State
Moved zoom into redux.  Moved a bunch of logic out of the canvas utility into the reducer to accomplish this and also simplified it a bit.  There were some files with 4 space indentation that were corrected as well.

Being able to undo/redo zoom is probably not too good, so that might necessitate a change to that system in the future.