Simplify Left Panel
Removed the two elements that were part of the left panel and replaced them with a single element full of buttons for use.  May need to work on the color picker to prevent it from becoming invisible.

The preview element still exists, but isn't tethered to anything and will need to be moved.

The material library was upgraded for what ultimately turned out to be no reason.